Move Spain & Region tests to a dedicated test module
refs #531
